+++ app/src/main/java/com/marcogn/hallofmemories/HallOfMemoriesApplication.kt
@@ -1,0 +1,7 @@
+package com.marcogn.hallofmemories
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class HallOfMemoriesApplication : Application()